general basegame feature
This adds a supported "-basegame" commandline arg, and new forms of the console "game" command to specify a basegame.

The set of info we actually care to specify is:
* an optional "base" gamedir to load assets from (at most one for now)
* a gamedir to load assets from
* whether to activate hipnotic or rogue code paths in the engine

Unfortunately we have legacy behaviors that make this messy, that we need to maintain for backwards compatibility:
* missionpack-specific "-hipnotic" and "-rogue" arguments that both set a basegame and activate code paths
* special "-quoth" argument that sets a basegame and activates the hipnotic code path
* "friendly" behavior when setting hipnotic/rogue/quoth gamedir that will also activate the relevant missionpack code path

So, hmm.

Here's one well-defined way to handle the "-basegame" commandline arg that still allows supporting those old behaviors:
* If you specify "-basegame" you must also specify "-game".
* You're not allowed to specify both "-basegame" and "-quoth".
* You can specify "-basegame" along with "-hipnotic"/"-rogue", but in that case the "-hipnotic"/"-rogue" argument only activates missionpack code.

We'll also have the "friendly" behavior where a hipnotic/rogue/quoth basegame can imply a missionpack code path to activate. We can only choose one missionpack code path though, so we need a prioritization for the various ways this can be specified or implied. Any actually specified missionpack argument has highest priority, followed by the basegame's implication, followed by the game's implication.

As for the console "game" command, it supports the same variety of configurations, while maintaining backward compat for the existing game command args. Described in detail in the comment header for COM_Game_f.

==========

Here's some examples of launching Quake to play mod directories based on Arcane Dimensions (in the "ad" directory) or Copper (in the "copper" directory):

  (other Quake arguments then) -game ad_jpqm5 -basegame ad 

  (other Quake arguments then) -game ctsj -basegame copper 

Or using the "game" command from the console:

  game ad_jpqm5 ad

  game ctsj copper